simulation_server/requirements.txt
# OsteoTwin Simulation Server dependencies
# FastAPI + ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.10.0

# 3D mesh processing
trimesh>=4.6.0
numpy>=2.0.0

# VTK for DICOM → mesh extraction
vtk>=9.4.0

# Medical imaging
pydicom>=3.0.0

# Collision detection (trimesh backends)
scipy>=1.14.0
rtree>=1.3.0

# STL export for 3D printing
numpy-stl>=3.1.0

# Inter-service auth
python-dotenv>=1.0.0
